Update sample-data-parser records fields to double

+++ algorithm-prototyping/sample-data-parser/src/main/java/edu/uga/devdogs/sampledataparser/records/Distances.java
@@ -13,5 +13,5 @@ import com.google.gson.annotations.SerializedName;
  */
 public record Distances(String[] buildings,
                         @SerializedName("distance_matrix")
-                        float[][] matrix) {
+                        double[][] matrix) {
 }
